* commit '632e244f6114611991122ca63af191ad823c3660':
  refactor：pom调整
  Refactor：fineio file使用数组缓存，减轻cache压力；使用隐式异常优化，减少判断；结构优化调整，test调整

+++ src/test/java/com/fineio/v3/file/impl/FileDemo.java
@@ -14,7 +14,6 @@ import org.junit.Before;
 import org.junit.Test;
 
 import java.util.stream.IntStream;
-import java.util.stream.LongStream;
 
 import static org.junit.Assert.assertEquals;
 
@@ -59,7 +58,7 @@ public class FileDemo {
     @Test
     public void testLong() {
         LongWriteFile longWriteFile = LongWriteFile.ofSync(key, connector);
-        LongStream.range(0, n).forEachOrdered(i -> longWriteFile.putLong(i, i));
+        IntStream.range(0, n).forEachOrdered(i -> longWriteFile.putLong(i, i));
         longWriteFile.close();
 
         LongReadFile longReadFile = new LongReadFile(key, connector);
@@ -71,7 +70,7 @@ public class FileDemo {
     @Test
     public void testDouble() {
         DoubleWriteFile doubleWriteFile = DoubleWriteFile.ofSync(key, connector);
-        LongStream.range(0, n).forEachOrdered(i -> doubleWriteFile.putDouble(i, i));
+        IntStream.range(0, n).forEachOrdered(i -> doubleWriteFile.putDouble(i, i));
         doubleWriteFile.close();
 
         DoubleReadFile doubleReadFile = new DoubleReadFile(key, connector);
